#5498: Move watcher out of llrt
