+++ tests/tt_metal/tt_metal/unit_tests_common/common/watcher_fixture.hpp
@@ -5,7 +5,7 @@
 #include <chrono>
 #include <thread>
 #include "common_fixture.hpp"
-#include "llrt/watcher.hpp"
+#include "impl/debug/watcher_server.hpp"
 
 // A version of CommonFixture with watcher enabled
 class WatcherFixture: public CommonFixture {
@@ -40,7 +40,7 @@ protected:
         tt::llrt::OptionsG.set_watcher_dump_all(false);
         tt::llrt::OptionsG.set_watcher_append(false);
         tt::llrt::OptionsG.set_test_mode_enabled(true);
-        tt::llrt::watcher_clear_log();
+        tt::watcher_clear_log();
 
         // Parent class initializes devices and any necessary flags
         CommonFixture::SetUp();
@@ -59,7 +59,7 @@ protected:
         tt::llrt::OptionsG.set_watcher_dump_all(watcher_previous_dump_all);
         tt::llrt::OptionsG.set_watcher_append(watcher_previous_append);
         tt::llrt::OptionsG.set_test_mode_enabled(test_mode_previous);
-        tt::llrt::watcher_server_clear_error_flag();
+        tt::watcher_server_clear_error_flag();
     }
 
     void RunTestOnDevice(
@@ -72,6 +72,6 @@ protected:
         CommonFixture::RunTestOnDevice(run_function_no_args, device);
         // Wait for a final watcher poll and then clear the log.
         std::this_thread::sleep_for(std::chrono::milliseconds(interval_ms));
-        tt::llrt::watcher_clear_log();
+        tt::watcher_clear_log();
     }
 };
