+++ tt_metal/impl/debug/watcher_server.hpp
@@ -4,10 +4,7 @@
 
 #pragma once
 
-#include "tt_cluster.hpp"
-
 namespace tt {
-namespace llrt {
 
 void watcher_init(int device_id,
                   std::function<CoreCoord ()>get_grid_size,
@@ -45,5 +42,4 @@ void watcher_clear_log();
 // Helper function to get the current watcher log file name/path
 string watcher_get_log_file_name();
 
-} // namespace llrt
 } // namespace tt
