+++ tt_metal/impl/device/device.cpp
@@ -8,6 +8,7 @@
 #include "tt_metal/third_party/tracy/public/tracy/Tracy.hpp"
 #include "tt_metal/detail/tt_metal.hpp"
 #include "impl/debug/dprint_server.hpp"
+#include "impl/debug/watcher_server.hpp"
 #include "tt_metal/third_party/umd/device/util.hpp"
 
 
@@ -548,7 +549,7 @@ bool Device::initialize(const std::vector<uint32_t>& l1_bank_remap) {
     }
 
     DprintServerAttach(this);
-    llrt::watcher_init(this->id(),
+    watcher_init(this->id(),
                        [&, this]() { return this->logical_grid_size(); },
                        [&, this](CoreCoord core) { return this->worker_core_from_logical_core(core); },
                        [&, this]() -> const std::unordered_set<CoreCoord> { return this->get_active_ethernet_cores(); },
@@ -557,7 +558,7 @@ bool Device::initialize(const std::vector<uint32_t>& l1_bank_remap) {
 
     this->initialize_and_launch_firmware();
 
-    llrt::watcher_attach(this, this->id(),
+    watcher_attach(this, this->id(),
                          [&, this]() { return this->logical_grid_size(); },
                          [&, this](CoreCoord core) { return this->worker_core_from_logical_core(core); },
                          [&, this]() -> const std::set<CoreCoord>& { return this->storage_only_cores(); },
@@ -587,7 +588,7 @@ bool Device::close() {
         TT_THROW("Cannot close device {} that has not been initialized!", this->id_);
     }
     this->deallocate_buffers();
-    llrt::watcher_detach(this);
+    watcher_detach(this);
     DprintServerDetach(this);
 
     // Assert worker cores
